requirements.txt
discord.py>=2.2.2
requests>=2.27.1